+++ CHANGELOG.md
@@ -6,7 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-- No changes yet.
+### Fixed
+- ([#2]): Patches with named imports now support matching and manipulating any
+  import.
+
+  [#2]: https://github.com/uber-go/gopatch/issues/2
 
 ## 0.0.2 - 2020-11-04
 ### Fixed